requirements.txt
beautifulsoup4==4.11.1
black==22.8.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
flake8==5.0.4
idna==3.3
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pycodestyle==2.9.1
pyflakes==2.5.0
requests==2.28.1
soupsieve==2.3.2.post1
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
supabase==0.5.8
python-dotenv==0.21.0
pytz==2022.2.1
python-crontab==2.6.0
